A poll() failure with EINTR does not mean a socket is closed.
This won't impact Synchronet as it has a separate signal handling
thread, but we still need to behave properly for processes that
don't.  I'm also saying that ENOMEM does not indicate a disconnection,
though it may be better to pretend it was disconnected...

+++ src/sbbs3/node.c
@@ -517,7 +517,7 @@ int main(int argc, char **argv)
 	sprintf(str,"%snode.exb",ctrl_dir);
 	nodeexb=sopen(str,O_RDWR|O_BINARY,SH_DENYNO);
 
-	sys_nodes=filelength(nodefile)/sizeof(node_t);
+	sys_nodes=(int)(filelength(nodefile)/sizeof(node_t));
 	if(!sys_nodes) {
 		printf("%s reflects 0 nodes!\n",str);
 		exit(1); }
